quick and dirty nav cleanup
