+++ www/components/Navbar/Item.js
@@ -2,12 +2,13 @@ import React, { PropTypes } from 'react';
 import Link from 'gatsby-link';
 import theme from '../../theme';
 
-export default function Item ({ title, url }) {
+export default function Item ({ depth, title, url }) {
+	console.log('depth', depth);
 	return (
 		<li css={styles.item} key={url}>
 			<Link
 				onlyActiveOnIndex
-				css={styles.link}
+				css={[styles.link, styles[`link__${depth}`]]}
 				activeStyle={styles.link__active}
 				to={url}
 			>
@@ -28,11 +29,11 @@ const styles = {
 	item: {
 		fontWeight: 300,
 		margin: '0 0 2px',
-		marginLeft: '20px',
+		// marginLeft: '20px',
 	},
 	link: {
-		borderBottomRightRadius: 3,
-		borderTopRightRadius: 3,
+		// borderBottomRightRadius: 3,
+		// borderTopRightRadius: 3,
 		color: 'white',
 		display: 'block',
 		padding: `0.5em 1rem`,
@@ -42,13 +43,14 @@ const styles = {
 		':hover': {
 			backgroundColor: 'rgba(255, 255, 255, 0.1)',
 		},
-		[theme.breakpoint.largeUp]: {
-			paddingLeft: `2rem`,
-			paddingRight: `2rem`,
-		},
 	},
 	link__active: {
 		backgroundColor: theme.color.blue,
 		opacity: 1,
+	},
+
+	// depth
+	link__2: {
+		paddingLeft: '2rem',
 	},
 };
