+++ www/components/Navbar/Menu.js
@@ -1,333 +1,52 @@
 import React from 'react';
-import theme from '../../theme';
+import { itemsShape, makeSection } from './utils';
+import { documentation, guides, welcome } from './data';
 
-import { itemsShape } from './utils';
-import Item from './Item';
-
-const welcome = {
-	section: 'Getting Started',
-	slug: '/getting-started',
-	items: [{
-		label: 'Getting Started',
-		slug: '',
-	}, {
-		label: 'Keystone Yeoman Generator',
-		slug: '/yo-generator',
-	}, {
-		section: 'Setting Up',
-		slug: '/setting-up',
-		items: [{
-			label: 'Part 1: Initial Setup',
-			slug: '/part-1',
-		}, {
-			label: 'Part 2 - Database Setup',
-			slug: '/part-2',
-		}, {
-			label: 'Part 3: Routing',
-			slug: '/part-3',
-		}, {
-			label: 'Part 4 -  Adding data to the database from a form',
-			slug: '/part-4',
-		}],
-	}],
-};
-const guides = {
-	section: 'Guides',
-	slug: '/guides',
-	items: [{
-		label: 'Upgrade Guide: 0.3 to 4.0',
-		slug: '/v-0-3-to-v-4-0-upgrade-guide',
-	}, {
-		label: 'API File and Image Uploads',
-		slug: '/api-file-image-uploads',
-	}, {
-		label: 'How to create a blog with Keystone',
-		slug: '/how-to-create-a-blog-with-keystone',
-	}, {
-		label: 'How to send emails with Keystone',
-		slug: '/how-to-send-emails',
-	}],
-};
-const documentation = {
-	section: 'Documentation',
-	slug: '/api',
-	items: [{
-		label: 'Documentation Index',
-		slug: '',
-	}, {
-		label: 'Configuration',
-		slug: '/configuration',
-	}, {
-		section: 'Database',
-		slug: '/database',
-		items: [{
-			label: 'Database Options',
-			slug: '',
-		}, {
-			label: 'Relationships',
-			slug: '/relationships',
-		}],
-	}, {
-		section: 'List',
-		slug: '/list',
-		items: [{
-			label: 'Creating a List',
-			slug: '',
-		}, {
-			label: 'Add',
-			slug: '/add',
-		}, {
-			label: 'Model',
-			slug: '/model',
-		}, {
-			label: 'List Construction Options',
-			slug: '/options',
-		}, {
-			label: 'Register',
-			slug: '/register',
-		}, {
-			label: 'Schema',
-			slug: '/schema',
-		}, {
-			label: 'Update Item',
-			slug: '/update-item',
-		}],
-	}, {
-		section: 'Methods',
-		slug: '/methods',
-		items: [{
-			label: 'Close Database connection',
-			slug: '/close-database-connection',
-		}, {
-			label: 'Create Items',
-			slug: '/create-items',
-		}, {
-			label: 'Create Router',
-			slug: '/create-router',
-		}, {
-			label: 'Get',
-			slug: '/get',
-		}, {
-			label: 'Import',
-			slug: '/import',
-		}, {
-			label: 'Importer',
-			slug: '/importer',
-		}, {
-			label: 'Init',
-			slug: '/init',
-		}, {
-			label: 'list',
-			slug: '/list',
-		}, {
-			label: 'Middleware',
-			slug: '/middleware',
-		}, {
-			label: 'Open Database Connection',
-			slug: '/open-database-connection',
-		}, {
-			label: 'Paginate',
-			slug: '/paginate',
-		}, {
-			label: 'Pre',
-			slug: '/pre',
-		}, {
-			label: 'Set',
-			slug: '/set',
-		}, {
-			label: 'Start',
-			slug: '/start',
-		}, {
-			label: 'Update Handler',
-			slug: '/update-handler',
-		}],
-	}, {
-		section: 'View',
-		slug: '/view',
-		items: [{
-			label: 'on',
-			slug: '/on',
-		}, {
-			label: 'query',
-			slug: '/query',
-		}, {
-			label: 'render',
-			slug: '/render',
-		}],
-	}, {
-		section: 'Field Types',
-		slug: '/field',
-		items: [{
-			label: 'Field Options',
-			slug: '/options',
-		}, {
-			label: 'AzureFile Field',
-			slug: '/azurefile',
-		}, {
-			label: 'Embedly Field',
-			slug: '/embedly',
-		}, {
-			label: 'File Field',
-			slug: '/file',
-		}, {
-			label: 'LocalFile Field',
-			slug: '/localfile',
-		}, {
-			label: 'Local Files',
-			slug: '/localfiles',
-		}, {
-			label: 'S3 File Field',
-			slug: '/s-3-file',
-		}, {
-			label: 'Boolean Field',
-			slug: '/boolean',
-		}, {
-			label: 'CloudinaryImage Field',
-			slug: '/cloudinaryimage',
-		}, {
-			label: 'CloudinaryImages Field',
-			slug: '/cloudinaryimages',
-		}, {
-			label: 'Code Field',
-			slug: '/code',
-		}, {
-			label: 'Date Field',
-			slug: '/date',
-		}, {
-			label: 'Color Field',
-			slug: '/color',
-		}, {
-			label: 'Datearray field',
-			slug: '/datearray',
-		}, {
-			label: 'DateTime Field',
-			slug: '/datetime',
-		}, {
-			label: 'Email Field',
-			slug: '/email',
-		}, {
-			label: 'GeoPoint Field',
-			slug: '/geopoint',
-		}, {
-			label: 'HTML Field',
-			slug: '/html',
-		}, {
-			label: 'Key Field',
-			slug: '/key',
-		}, {
-			label: 'Location Field',
-			slug: '/location',
-		}, {
-			label: 'Markdown Field',
-			slug: '/markdown',
-		}, {
-			label: 'Money Field',
-			slug: '/money',
-		}, {
-			label: 'Name Field',
-			slug: '/name',
-		}, {
-			label: 'Number Field',
-			slug: '/number',
-		}, {
-			label: 'NumberArray field',
-			slug: '/numberarray',
-		}, {
-			label: 'Password Field',
-			slug: '/password',
-		}, {
-			label: 'Relationship Field',
-			slug: '/relationship',
-		}, {
-			label: 'Select Field',
-			slug: '/select',
-		}, {
-			label: 'Text Field',
-			slug: '/text',
-		}, {
-			label: 'Textarea Field',
-			slug: '/textarea',
-		}, {
-			label: 'Textarray Field',
-			slug: '/textarray',
-		}, {
-			label: 'URL Field',
-			slug: '/url',
-		}],
-	}],
-};
-
-const menu = [
+const sections = [
 	welcome,
 	guides,
 	documentation,
 ];
 
-function makeSection (currentPath, layer, level) {
-	return layer.map((section, idx) => {
-		const styleChoice = level === 1 ? styles.section : styles.subsection;
-
-		return (
-			<ul key={idx} css={styles[`menu${level}`]}>
-				<li css={styleChoice}>{section.section}</li>
-				{section.items ? section.items.map(function (item) {
-					const newPath = currentPath + section.slug;
-					if (item.items) {
-						return makeSection(newPath, [item], (level + 1));
-					}
-					return (
-						<Item
-							key={item.slug}
-							title={item.label}
-							url={newPath + item.slug}
-						/>
-					);
-				}) : null}
-			</ul>
-		);
-	});
-}
-
-export default function Menu ({ items }) {
-	const list = makeSection('', menu, 1);
-
-	return <nav>{list}</nav>;
+export default function Menu () {
+	// let items = [];
+	// sections.forEach(s => {
+	// 	items.push(<h2 key={s.section}>{s.section}</h2>);
+	//
+	// 	s.items.forEach(i => {
+	//
+	// 	});
+	// });
+	// console.table(sections);
+	return <nav>{makeSection('', sections, 1)}</nav>;
 };
+
+// return layer.map((section, idx) => {
+// 	const styleChoice = level === 1 ? styles.section : styles.subsection;
+//
+// 	return (
+// 		<ul key={idx} css={[styles.menu, styles[`menu_depth_${level}`]]}>
+// 			<li css={styleChoice}>{section.section}</li>
+//
+// 			{section.items ? section.items.map(function (item) {
+// 				const newPath = currentPath + section.slug;
+//
+// 				if (item.items) {
+// 					return makeSection(newPath, [item], (level + 1));
+// 				}
+//
+// 				return (
+// 					<Item
+// 						key={item.slug}
+// 						title={item.label}
+// 						url={newPath + item.slug}
+// 					/>
+// 				);
+// 			}) : null}
+// 		</ul>
+// 	);
+// });
 
 Menu.propTypes = {
 	items: itemsShape.isRequired,
-};
-
-/* eslint quote-props: ["error", "as-needed"] */
-const styles = {
-	menu1: {
-		listStyle: 'none',
-		margin: 0,
-		paddingRight: '1em',
-	},
-	menu2: {
-		listStyle: 'none',
-		marginTop: 0,
-		marginBottom: 0,
-		marginLeft: '20px',
-		paddingRight: '1em',
-	},
-	section: {
-		fontSize: '1.7em',
-		marginTop: '1.8em',
-		opacity: 0.6,
-		padding: `0 1rem`,
-		textTransform: 'uppercase',
-		[theme.breakpoint.largeUp]: {
-			paddingLeft: `2rem`,
-			paddingRight: `2rem`,
-		},
-	},
-	subsection: {
-		fontSize: '1.3em',
-		opacity: 0.6,
-		padding: `0 1rem`,
-		marginLeft: '20px',
-		textTransform: 'uppercase',
-	},
 };
