+++ www/components/Navbar/utils/index.js
@@ -1,9 +1,15 @@
 import { PropTypes } from 'react';
+import makeSection from './makeSection';
 
-export const itemsShape = PropTypes.arrayOf(PropTypes.shape({
+const itemsShape = PropTypes.arrayOf(PropTypes.shape({
 	section: PropTypes.string,
 	items: PropTypes.arrayOf(PropTypes.shape({
 		label: PropTypes.string,
 		slug: PropTypes.string,
 	})),
 }));
+
+export default {
+	itemsShape,
+	makeSection,
+};
