+++ www/css/styles.css
@@ -1,18 +1,25 @@
-body {
-  font-size: 13px;
-  line-height: 1.4;
+html {
+	font-size: 16px;
 }
-code, [data-primitive-type]::after {
-  border-radius: 2px;
-  opacity: 0.8;
-  padding: 0 4px 1px;
-  border-right: 3px solid white;
-}
-h4>code {
-  background-color: lightpink;
+@media (max-width: 1024px) { html { font-size: 14px; } }
+@media (max-width: 768px) { html { font-size: 13px; } }
+
+h4 > code {
+	background-color: #f9f2f4;
+	/*border-radius: 4px;*/
+	color: #c7254e;
+	font-size: 90%;
+	line-height: 1;
+	/*padding: 2px 4px;*/
 }
 [data-primitive-type]::after {
-  background-color: lightskyblue;
-  content: attr(data-primitive-type);
-  font-family: monospace;
+	background: #f9f9f9;
+	border-radius: 3px;
+	color: #0e487a;
+	content: attr(data-primitive-type);
+	font-family: Monaco, Menlo, Courier, monospace;
+	line-height: 1;
+	margin-left: 0.5em;
+	padding: 2px 4px;
+	white-space: nowrap;
 }
